+++ src/Pages/Contacts.jsx
@@ -1,11 +1,10 @@
 ﻿import React from 'react';
-import s from './contacts.module.css'
 
 const Contacts = () => {
     return (
-        <main id={s.contact__page}>
+        <div>
 
-        </main>
+        </div>
     );
 };
