--- src/Pages/ContactsPage/contacts.module.css
@@ -1,1 +1,0 @@
-﻿
