+++ src/Pages/Entry.jsx
@@ -1,11 +1,10 @@
 ﻿import React from 'react';
-import s from './entry.module.css'
 
 const Entry = () => {
     return (
-        <main id={s.entry__page}>
+        <div>
             huy
-        </main>
+        </div>
     );
 };
