--- src/Pages/Entry/entry.module.css
@@ -1,1 +1,0 @@
-﻿
