--- src/Pages/Home/home.module.css
@@ -1,1 +1,0 @@
-﻿
