+++ src/components/HomeComponents/sectionServices/sectionServices.module.css
@@ -62,6 +62,6 @@
   gap: 52px;
 }
 
-#services > .learn__blocks > .block:hover{
-  background-color: #4200FF;
+#services > .learn__blocks > .block:hover {
+  background-color: #4200ff;
 }
