+++ src/components/UI/SideBar/Sidebar.module.scss
@@ -1,8 +1,7 @@
-﻿/* .App {
-  font-family: sans-serif;
-  text-align: center;
-} */
-
+﻿// .App {
+//   font-family: sans-serif;
+//   text-align: center;
+// }
 
 #App {
   font-family: sans-serif;
@@ -64,3 +63,5 @@
 .bm-overlay {
   background: rgba(0, 0, 0, 0.3);
 }
+
+
